KAN-87 refactor: footerData.ts 파일 위치 변경 (footer → data)
